overhaul api points and added security with access tokens

+++ src/main/java/de/leonlatsch/oliviabackend/service/UserService.java
@@ -2,15 +2,14 @@ package de.leonlatsch.oliviabackend.service;
 
 import de.leonlatsch.oliviabackend.dto.AuthResponse;
 import de.leonlatsch.oliviabackend.dto.ProfilePicDTO;
+import de.leonlatsch.oliviabackend.dto.PublicUserDTO;
 import de.leonlatsch.oliviabackend.dto.UserDTO;
 import de.leonlatsch.oliviabackend.entity.AccessToken;
 import de.leonlatsch.oliviabackend.entity.User;
 import de.leonlatsch.oliviabackend.repository.AccessTokenRepository;
 import de.leonlatsch.oliviabackend.repository.UserRepository;
+import de.leonlatsch.oliviabackend.util.*;
 import de.leonlatsch.oliviabackend.util.Base64;
-import de.leonlatsch.oliviabackend.util.CommonUtils;
-import de.leonlatsch.oliviabackend.util.ImageHelper;
-import de.leonlatsch.oliviabackend.util.DatabaseMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,27 +31,31 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private AccessTokenService accessTokenService;
+
     private DatabaseMapper mapper = DatabaseMapper.getInstance();
 
-    public List<UserDTO> getAllUsers() {
+    public List<UserDTO> getAllUsers(String accessToken) {
+        if (!AdminManager.getAdminAccessToken().equals(accessToken)) {
+            return null;
+        }
         List<UserDTO> list = mapToTransferObjects(userRepository.findAll());
 
         for (UserDTO user : list) {
-            rmPic(user);
+            rmProfilePic(user);
         }
 
         return list;
     }
 
-    public UserDTO getUserByUid(int uid) {
+    public UserDTO get(String accessToken) {
+        int uid = accessTokenService.getUserForToken(accessToken);
+        if (uid == -1) {
+            return null;
+        }
         Optional<User> user = userRepository.findById(uid);
-        rmPic(user);
-        return user.isPresent() ? mapper.mapToTransferObject(user.get()) : null;
-    }
-
-    public UserDTO getUserByEmail(String email) {
-        Optional<User> user = userRepository.findByEmail(email);
-        rmPic(user);
+        rmProfilePic(user);
         return user.isPresent() ? mapper.mapToTransferObject(user.get()) : null;
     }
 
@@ -98,8 +101,13 @@ public class UserService {
         return response;
     }
 
-    public String deleteUser(int uid) {
+    public String deleteUser(String accessToken) {
+        int uid = accessTokenService.getUserForToken(accessToken);
+        if (uid == -1) {
+            return ERROR;
+        }
         userRepository.deleteById(uid);
+        accessTokenService.disableAccessToken(accessToken);
         return OK;
     }
 
@@ -113,13 +121,15 @@ public class UserService {
         return user.isPresent() ? TAKEN : FREE;
     }
 
-    public String updateUser(UserDTO userDTO) {
-        if (userDTO.getUid() < 10000000) {
-            return "UID_IS_NULL";
+    public String updateUser(String accessToken, UserDTO userDTO) {
+        int uid = accessTokenService.getUserForToken(accessToken);
+
+        if (uid == -1) {
+            return ERROR;
         }
         User user = mapper.mapToEntity(userDTO);
 
-        Optional<User> dbUser = userRepository.findById(user.getUid());
+        Optional<User> dbUser = userRepository.findById(uid);
         if (dbUser.isPresent()) {
             if (user.getUsername() == null) {
                 user.setUsername(dbUser.get().getUsername());
@@ -143,32 +153,34 @@ public class UserService {
         }
     }
 
-    public List<UserDTO> getUserByUsername(String username) {
+    public List<PublicUserDTO> search(String username) {
         List<User> users = userRepository.findByUsernameContaining(username);
         for (User user : users) {
-            rmPic(user);
+            rmProfilePic(user);
         }
-        return mapToTransferObjects(users);
+        return mapToPublicUsers(mapToTransferObjects(users));
     }
 
-    public List<UserDTO> getUserTop100(String username) {
+    public List<PublicUserDTO> searchTop100(String username) {
         List<User> users = userRepository.findTop100ByUsernameContaining(username);
         for (User user : users) {
-            rmPic(user);
+            rmProfilePic(user);
         }
 
-        return mapToTransferObjects(users);
+        return mapToPublicUsers(mapToTransferObjects(users));
     }
 
     public AuthResponse authUserByEmail(String email, String hash) {
         Optional<User> user = userRepository.findByEmail(email);
+        AuthResponse response = new AuthResponse();
 
         if (hash == null || !user.isPresent()) {
-            return null;
+            response.setMessage(UNAUTORIZED);
+            response.setSuccess(false);
+            return response;
         }
 
-        AuthResponse response = new AuthResponse();
-        String token = loadAccessToken(user.get().getUid());
+        String token = accessTokenService.getTokenForUser(user.get().getUid());
         if (user.get().getPassword().equals(hash) && token != null) {
             response.setMessage(AUTORIZED);
             response.setAccessToken(token);
@@ -182,13 +194,11 @@ public class UserService {
         }
     }
 
-    private String loadAccessToken(int uid) {
-        Optional<AccessToken> token = accessTokenRepository.findByUid(uid);
-
-        return token.isPresent() && token.get().isValid() ? token.get().getToken() : null;
-    }
-
-    public ProfilePicDTO loadProfilePic(int uid) {
+    public ProfilePicDTO loadProfilePic(String accessToken) {
+        int uid = accessTokenService.getUserForToken(accessToken);
+        if (uid == -1) {
+            return null;
+        }
         ProfilePicDTO profilePicDto = new ProfilePicDTO();
         Optional<User> user = userRepository.findById(uid);
         if (user.isPresent()) {
@@ -198,6 +208,18 @@ public class UserService {
         }
 
         return profilePicDto;
+    }
+
+    private List<PublicUserDTO> mapToPublicUsers(Collection<UserDTO> users) {
+        if (users == null) {
+            return null;
+        }
+
+        List<PublicUserDTO> publicUsers = new ArrayList<>();
+        for (UserDTO dto : users) {
+            publicUsers.add(mapper.mapToPublicUser(dto));
+        }
+        return publicUsers;
     }
 
     private List<UserDTO> mapToTransferObjects(Collection<User> entities) {
@@ -211,17 +233,17 @@ public class UserService {
         return transferObjects;
     }
 
-    private void rmPic(UserDTO dto) {
+    private void rmProfilePic(UserDTO dto) {
         dto.setProfilePic(null);
     }
 
-    private void rmPic(Optional<User> dto) {
+    private void rmProfilePic(Optional<User> dto) {
         if (dto.isPresent()) {
             dto.get().setProfilePic(null);
         }
     }
 
-    private void rmPic(User user) {
+    private void rmProfilePic(User user) {
         user.setProfilePic(null);
     }
 }
